requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
PyGithub==2.1.1
google-generativeai==0.3.2
pydantic==2.5.3
pydantic-settings==2.1.0